requirements.txt
Django==2.2.1
djangorestframework==3.9.3
pytz==2019.1
sqlparse==0.3.0
psycopg2-binary==2.8.2
django-filter==2.1.0
